Add a KDE notification when wallet sync to disk fails
BUG: 226742
FIXED-IN: 4.12

This does not really fixes the bug, as I'm not able to reproduce it
and never happened on my system. However, KSaveFile's logic may fail
and kwalletd did not took that into account. When such a failure
occurred, kwalletd silently ignored the errors. This fixes that and
adds a system notification to let the user know that something is
wrong with the disk (disk full, r/W error, etc.)