Remove the autoclose tests because they don't really test that function
The autoclose test, for instance, opens the wallet then quits. Instead
it should setup a timer whose duration is longer than the autoclose delay
configured in kwalletd. After that delay, it should test the wallet
really autoclosed in kwalletd. For the test to work, kwalletd should be
configured to do aucoclose and the timeout should be set to be shorter
that the delay hard-coded in the test.

Adding a TODO with this test and eventually implementing later, if secret
service will allow some time for this (any help appreciated)