Partially revert 321d604f451b0026f55519d4ff695c185a925f00
The problem that the commit tried to prevent was actually a different
problem in kwalletmanager (and virtually all apps opening a wallet).

The core of the problem is clients sending wrong WId to KWallet by
invoking winId() inside the widget's code where it's called. That will
pass only that widget's WId, not the WId of an actual window the widget
belongs to. Therefore the check for active window would always fail in
those cases -->

if (w != KWindowSystem::activeWindow() ... )

because the widget, whose WId was passed, can never be active window
(unless that widget actually /is/ a window, in which case all is good).

What clients should do is either pass window()->winId() or pass
effectiveWinId() which always returns the window WId.

I'll also make a commit to kwalletmanager to fix this.

Maybe KWallet's docs should be updated including this info?

CCMAIL: kde@rusu.info
CCMAIL: teo@kde.org