'walletListDirty' signal should be emitted if the 'kwalletd' directory is deleted
If it is not emitted, kwalletmanager would continue to show the wallets even though
it doesnt exist on the disk anymore. These would be removed from the list only when
the wallet is closed. If the user had added any new data to the wallets, it would be
lost once the wallet is closed.

REVIEW: 120387