Fix the CBC encryption algorith
The CBC algorithm was not enrcrypting in CBC but in ECB, despite it being
named CBC. This problem was found by Itay Duvdevani who let us know it on the
security mailing list. His mail eventually got forwarded to me and here is the
fix. I also fixed the test program which was incorrectly checking the expected
bytes.