kwalletd: Fix FILE* leak.
Fix a leak of a `FILE*` in kwalletd, spotted by Coverity (CID 1335090), and a
couple of minor fixes/cleanups since I was there.

1. No need to memset a `char[]`, you can let compiler default-initialize it
(and likely even be able to move the whole array into .bss).

2. The chop after `fgets` can easily be confused for a null terminator check,
and `fgets` isn't required to return a newline anyways, so add sanity checking
(which helps document better at the same time).

REVIEW:126410