KWallet::openWallet(Synchronous) : don't time out after 25 seconds.
The default DBus timeout is 25 seconds, which means that if the user went to
get a cup of tea during session startup, when they come back they get prompted
with all sorts of additional non-kwallet password requests due to all kwallet
requests having timed out.

I added setTimeout in QDBusAbstractInterface in Qt 4.8 for things like this.

Testcase:
 eval `dbus-launch`
 export SSH_ASKPASS=$KDEDIR/bin/ksshaskpass
 ssh-add < /dev/null
(but the same happens with the IMAP resource etc.)

REVIEW: 127121