Move kwalletd initialization earlier
In the previous code (before removing the kde4 migration agent) it was
started inconditionally, while after the change it was gated behind an
if.

However this does not take into account GPG wallets, which, as far as
I understand from the code, do not produce a password hash.

Moving initialization out of the if restores GPG functionality

Stefan, please check if everything is correct.

CCMAIL: stefan.bruens@rwth-aachen.de

BUG: 410020