Revert readEntryList() to use QRegExp::Wildcard
Summary:
The issue with QRegularExpression::wildcardToRegularExpression() is
that it's more strict in its interpretation of a wildcard, i.e. it
makes it correctly work with file globbing patterns, meaning it would
return this pattern "\\A(?:[^/]*)\\z" when called with "*", it excludes
"/" which is a forbidden character in a filename, and it anchors the pattern
(which is what users of QRegExp did by using QRegExp::exactMatch()).
readEntryList() uses the regex to match against entries of the form:
WEBSITE / USERNAME, which includes a "/".

Test Plan:
- Install falkon (and falkon-kde for kwallet integration, or whatever
  your distro calls it) and log in to a website with a user name and a
  password and select to remember it when asked to.
- Restart falkon and see if the login data is shown

with this patch it should show the login data.

Reviewers: #frameworks, dfaure, apol

Reviewed By: dfaure

Subscribers: blaze, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26734