[KWallet] Port last usage of QRegExp to QRegularExpression
Summary:
QRegularExpression::wildcardToRegularExpression() mainly handles file
pattern globbing (e.g. "*.txt") which means it doesn't allow "/" in
the file name (which is technically correct); but we have to subvert
it because the keys in kwallet are in the form "foobar.com/<User name>"
which does have a "/" in it. It's a hack, but it works here because
in reality, there're no actual files/folders in kwallet, it's just
data structures and "/" is an acceptable separator.

Test Plan:
- make && ctest
- The password dialog in Falkon still wors

Reviewers: #frameworks, dfaure, blaze

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T12279

Differential Revision: https://phabricator.kde.org/D28880