Introduce three new methods that return all "entries" in a folder
Summary:
As was discussed in https://phabricator.kde.org/D28880 with dfaure, it
seems that all users (only two, KWalletManager and Falkon) of the current
read{Entry,Map,Passowrd}list() methods basically use "*" as a wildcard
to get a list of all the "entries" in a folder. Therefore it makes sense
to introduce new methods that do just that, this gets rid of regular
expression usage for matching a certain pattern. This fits with how .e.g.
KWalletManager is using readEntryList(), to get a list of all the entries
to fill a list view with them, then the "matching" is done with a QSFPM-like
class.

Test Plan:
- make && ctest
- Ported Falkon locally, and it seems to work