Un-overload OrgKdeKWalletInterface::walletClosed(int) signal
Add a KF6 TODO to remove the walletClosed(int) signal, but leave the signal
so as not to break apps listening on the DBus signal, because the breakage
would probably just make it fail silently.

Add a new signal, walletClosedId(int), this makes it slightly easier to
use the signal in QObject::connect() calls, without needing to use e.g.
QOverload, which is hard to write and read.

Add DBus deprecation annotation to the DBus interface XML file.

KF6 task: https://phabricator.kde.org/T13986