Drop obsolete KSecretsService integration
This code integrates with the out-of-tree and long-abandoned KSecretsService API, but it has been disabled
at build time for years.

The intention seems to use whatever secrets provider provides org.freedesktop.Secrets through the KWallet API.

The idea is certainly interesting, but given it has been defunct "forever" there's little point in keeping this code.

A proper solution would involve using QtKeyChain as an interface library