Register both the 5 and 6 versioned D-Bus service and path
This allows for backward compatibility with consumers looking for
the version 5 names, while also working for consumers that use
KDE_SESSION_VERSION for this (like Chrome). It also allows to
fix consumers (such as QtKeychain) towards the final interface for
6 rather than using temporary workarounds for the current mix.

This still needs a full rename of the binary and the D-Bus service
file etc, in particular this will still not start the 6 kwalletd if
the 5 one is already running due to having the same name. A pure 6
setup doesn't have that problem.

The consumer side also still needs to be adjusted.