Fix reply type in portal implementation
The old implementation returned only a single int ("i") when the spec defines a reply type of "ua{sv}", an unsigned int as well as a dict mapping strings to variants.

This fixes that, making it so the method returns something xdg-desktop-portal expects and so hopefully works properly.

We don't have anything actually useful to put into the return value (the actual result is transferred via file descriptor).