Add secret service bridge
The KWallet daemon is replaced by a new daemon which registers
itself on the KWallet dbus service name and exposes all its old api.

But instead of using directly the kwallet backend to store secrets,
it proxies the api requests to a SecretService daemon, offering a
compatibility layer for old kwallet-using applications.
It is to be seen as legacy support and migration aid

This daemon uses the same metadata format of QtKeychain, so
when the application will migrate to it, a further data
migration shouldn't be necessary

The old KWalled daemon is still there: and is called KSecretd.
It doesn't expose the kwallet dbus api animore, but only the
SecretService api. It's used as the default secretservice provider
for the new proxy kwalletd and it's launched by it unless is
explicitly configured to not do so

CCBUG:491280

removeqtkeychain dependency