Use desktop file to launch kwalletmanager
kstart puts the started process in a cgroup, which is good
and what we want. However when started like this the cgroup
naming doesn't follow the standard pattern to encode the desktop
file name so tools like plasma-systemmonitor can't resolve
the desktop file from it.

Instead of giving the raw command ask kstart to start the
corresponding desktop file, which makes this work