Fix compile kwallet using kde-builder
The .kde-ci.yml defines a dependency to third-party/gpgme what will
result in gpgme being compiled when compiling kwallet rather then
using the distros gpgme.

The problem with that is that the newest gpgme requires a newer
libgpg-error then my distributor ships. That means an additional
dependency of kwallet to third-party/libgpg-error would be
needed to use the newer libgpg-error. That in turn would also need
a dependency to third-party/gpgmepp.

By removing the existing entry to third-party/gpgme all that is not
needed and the distributors gpgme+libgpg-error+gpgmepp will be used.

BUG:485564